♻️ match.service 도메인별 분할 (748 → 60+163+224+64+88+31줄)
- match.service.ts (60줄): 컨트롤러 호환성을 위한 파사드
- match-post.service.ts (163줄): 포스트 CRUD
- match-application.service.ts (224줄): 신청 CRUD + 채팅룸 생성
  ※ 200줄 룰 12% 초과, 신청 라이프사이클이 한 도메인이라 더 쪼개면 인위적
- match.formatter.ts: formatMatchPost/formatMatchApplication 헬퍼 추출
- match.module: 3개 서비스 모두 providers 등록
