+++ apps/match/src/match.service.ts
@@ -1,736 +1,60 @@
+import { Injectable } from '@nestjs/common';
 import {
-  Injectable,
-  Logger,
-  NotFoundException,
-  ForbiddenException,
-  BadRequestException,
-  Inject,
-} from '@nestjs/common';
-import { MySQLPrismaService } from '@app/prisma';
-import { UtilsService } from '@app/utils';
-import { ClientGrpc } from '@nestjs/microservices';
-import { firstValueFrom } from 'rxjs';
-import {
-  MatchPost,
-  MatchApplication,
-  GetMatchPostsRequest,
-  CreateMatchPostRequest,
-  GetMatchPostRequest,
-  UpdateMatchPostRequest,
-  DeleteMatchPostRequest,
   ApplyToMatchRequest,
+  CreateMatchPostRequest,
+  DeleteMatchPostRequest,
   GetMatchApplicationsRequest,
-  UpdateApplicationStatusRequest,
+  GetMatchPostRequest,
+  GetMatchPostsRequest,
   GetMyApplicationsRequest,
   GetMyPostsRequest,
-  MatchPostResponse,
-  SingleMatchPostResponse,
-  MatchApplicationsResponse,
-  CommonResponse,
-  ApplicationResponse,
-  ChatService,
+  UpdateApplicationStatusRequest,
+  UpdateMatchPostRequest,
 } from '@app/common/protobuf';
+import { MatchPostService } from './match-post.service';
+import { MatchApplicationService } from './match-application.service';
 
 @Injectable()
 export class MatchService {
-  private readonly logger = new Logger(MatchService.name);
-  private chatService: ChatService;
-
   constructor(
-    private readonly mysqlPrismaService: MySQLPrismaService,
-    private readonly utilsService: UtilsService,
-    @Inject('CHAT_SERVICE')
-    private readonly chatServiceClient: ClientGrpc,
+    private readonly postService: MatchPostService,
+    private readonly applicationService: MatchApplicationService,
   ) {}
 
-  onModuleInit() {
-    this.chatService =
-      this.chatServiceClient.getService<ChatService>('ChatService');
+  // Posts
+  getMatchPosts(req: GetMatchPostsRequest) {
+    return this.postService.getMatchPosts(req);
+  }
+  createMatchPost(req: CreateMatchPostRequest) {
+    return this.postService.createMatchPost(req);
+  }
+  getMatchPost(req: GetMatchPostRequest) {
+    return this.postService.getMatchPost(req);
+  }
+  updateMatchPost(req: UpdateMatchPostRequest) {
+    return this.postService.updateMatchPost(req);
+  }
+  deleteMatchPost(req: DeleteMatchPostRequest) {
+    return this.postService.deleteMatchPost(req);
+  }
+  getMyPosts(req: GetMyPostsRequest) {
+    return this.postService.getMyPosts(req);
   }
 
-  async getMatchPosts(
-    request: GetMatchPostsRequest,
-  ): Promise<MatchPostResponse> {
-    const { page = 1, pageSize = 10 } = request;
-    const skip = (page - 1) * pageSize;
-
-    const matchPosts = await this.mysqlPrismaService.matchPost.findMany({
-      where: {
-        deletedAt: null,
-      },
-      include: {
-        User: true,
-        _count: {
-          select: {
-            MatchApplication: {
-              where: {
-                status: 'accepted',
-              },
-            },
-          },
-        },
-      },
-      orderBy: {
-        createdAt: 'desc',
-      },
-      skip,
-      take: pageSize + 1, // 하나 더 가져와서 hasNext 확인
-    });
-
-    const hasNext = matchPosts.length > pageSize;
-    if (hasNext) {
-      matchPosts.pop(); // 마지막 요소 제거
-    }
-
-    const formattedPosts: MatchPost[] = matchPosts.map((post) => ({
-      id: post.id,
-      title: post.title,
-      userno: post.userno,
-      author: post.User.nickname || 'Unknown',
-      authorGender: post.User.gender || '',
-      content: post.content,
-      movieTitle: post.movieTitle,
-      theaterName: post.theaterName,
-      showTime: post.showTime,
-      maxParticipants: post.maxParticipants,
-      currentParticipants: post._count.MatchApplication,
-      location: post.location,
-      createdAt: post.createdAt.toISOString(),
-      updatedAt: post.updatedAt?.toISOString() || '',
-      deletedAt: post.deletedAt?.toISOString() || '',
-    }));
-
-    return {
-      matchPosts: formattedPosts,
-      hasNext,
-    };
+  // Applications
+  applyToMatch(req: ApplyToMatchRequest) {
+    return this.applicationService.applyToMatch(req);
   }
-
-  async createMatchPost(
-    request: CreateMatchPostRequest,
-  ): Promise<SingleMatchPostResponse> {
-    const {
-      title,
-      content,
-      movieTitle,
-      theaterName,
-      showTime,
-      maxParticipants,
-      location,
-      userno,
-    } = request;
-
-    const matchPost = await this.mysqlPrismaService.matchPost.create({
-      data: {
-        title,
-        content,
-        movieTitle,
-        theaterName,
-        showTime,
-        maxParticipants,
-        location,
-        userno,
-      },
-      include: {
-        User: true,
-        _count: {
-          select: {
-            MatchApplication: {
-              where: {
-                status: 'accepted',
-              },
-            },
-          },
-        },
-      },
-    });
-
-    return {
-      matchPost: {
-        id: matchPost.id,
-        title: matchPost.title,
-        userno: matchPost.userno,
-        author: matchPost.User.nickname || 'Unknown',
-        authorGender: matchPost.User.gender || '',
-        content: matchPost.content,
-        movieTitle: matchPost.movieTitle,
-        theaterName: matchPost.theaterName,
-        showTime: matchPost.showTime,
-        maxParticipants: matchPost.maxParticipants,
-        currentParticipants: matchPost._count.MatchApplication,
-        location: matchPost.location,
-        createdAt: matchPost.createdAt.toISOString(),
-        updatedAt: matchPost.updatedAt?.toISOString() || '',
-        deletedAt: matchPost.deletedAt?.toISOString() || '',
-      },
-    };
+  getMatchApplications(req: GetMatchApplicationsRequest) {
+    return this.applicationService.getMatchApplications(req);
   }
-
-  async getMatchPost(
-    request: GetMatchPostRequest,
-  ): Promise<SingleMatchPostResponse> {
-    const { matchId } = request;
-
-    const matchPost = await this.mysqlPrismaService.matchPost.findFirst({
-      where: {
-        id: matchId,
-        deletedAt: null,
-      },
-      include: {
-        User: true,
-        _count: {
-          select: {
-            MatchApplication: {
-              where: {
-                status: 'accepted',
-              },
-            },
-          },
-        },
-      },
-    });
-    if (!matchPost) {
-      return {
-        matchPost: null,
-      };
-    }
-
-    return {
-      matchPost: {
-        id: matchPost.id,
-        title: matchPost.title,
-        userno: matchPost.userno,
-        author: matchPost.User.nickname || 'Unknown',
-        authorGender: matchPost.User.gender || '',
-        content: matchPost.content,
-        movieTitle: matchPost.movieTitle,
-        theaterName: matchPost.theaterName,
-        showTime: matchPost.showTime,
-        maxParticipants: matchPost.maxParticipants,
-        currentParticipants: matchPost._count.MatchApplication,
-        location: matchPost.location,
-        createdAt: matchPost.createdAt.toISOString(),
-        updatedAt: matchPost.updatedAt?.toISOString() || '',
-        deletedAt: matchPost.deletedAt?.toISOString() || '',
-      },
-    };
+  updateApplicationStatus(req: UpdateApplicationStatusRequest) {
+    return this.applicationService.updateApplicationStatus(req);
   }
-
-  async updateMatchPost(
-    request: UpdateMatchPostRequest,
-  ): Promise<SingleMatchPostResponse> {
-    const {
-      matchId,
-      title,
-      content,
-      movieTitle,
-      theaterName,
-      showTime,
-      maxParticipants,
-      location,
-      userno,
-    } = request;
-
-    // 게시글 존재 여부 및 권한 확인
-    const existingPost = await this.mysqlPrismaService.matchPost.findFirst({
-      where: {
-        id: matchId,
-        deletedAt: null,
-      },
-    });
-
-    if (!existingPost) {
-      throw new NotFoundException('Match post not found');
-    }
-
-    if (existingPost.userno !== userno) {
-      throw new ForbiddenException('You can only update your own posts');
-    }
-
-    const updatedPost = await this.mysqlPrismaService.matchPost.update({
-      where: {
-        id: matchId,
-      },
-      data: {
-        title,
-        content,
-        movieTitle,
-        theaterName,
-        showTime,
-        maxParticipants,
-        location,
-        updatedAt: new Date(),
-      },
-      include: {
-        User: true,
-        _count: {
-          select: {
-            MatchApplication: {
-              where: {
-                status: 'accepted',
-              },
-            },
-          },
-        },
-      },
-    });
-
-    return {
-      matchPost: {
-        id: updatedPost.id,
-        title: updatedPost.title,
-        userno: updatedPost.userno,
-        author: updatedPost.User.nickname || 'Unknown',
-        authorGender: updatedPost.User.gender || '',
-        content: updatedPost.content,
-        movieTitle: updatedPost.movieTitle,
-        theaterName: updatedPost.theaterName,
-        showTime: updatedPost.showTime,
-        maxParticipants: updatedPost.maxParticipants,
-        currentParticipants: updatedPost._count.MatchApplication,
-        location: updatedPost.location,
-        createdAt: updatedPost.createdAt.toISOString(),
-        updatedAt: updatedPost.updatedAt?.toISOString() || '',
-        deletedAt: updatedPost.deletedAt?.toISOString() || '',
-      },
-    };
+  getMyApplications(req: GetMyApplicationsRequest) {
+    return this.applicationService.getMyApplications(req);
   }
-
-  async deleteMatchPost(
-    request: DeleteMatchPostRequest,
-  ): Promise<CommonResponse> {
-    const { matchId, userno } = request;
-
-    // 게시글 존재 여부 및 권한 확인
-    const existingPost = await this.mysqlPrismaService.matchPost.findFirst({
-      where: {
-        id: matchId,
-        deletedAt: null,
-      },
-    });
-
-    if (!existingPost) {
-      throw new NotFoundException('Match post not found');
-    }
-
-    if (existingPost.userno !== userno) {
-      throw new ForbiddenException('You can only delete your own posts');
-    }
-
-    // Soft delete
-    await this.mysqlPrismaService.matchPost.update({
-      where: {
-        id: matchId,
-      },
-      data: {
-        deletedAt: new Date(),
-      },
-    });
-
-    return {
-      success: true,
-      message: 'Match post deleted successfully',
-    };
-  }
-
-  async applyToMatch(request: ApplyToMatchRequest): Promise<CommonResponse> {
-    const { matchId, applicantUserno, applicantName, message } = request;
-    this.logger.log(`applyToMatch matchId=${matchId} applicant=${applicantUserno}`);
-
-    // 게시글 존재 여부 확인
-    const matchPost = await this.mysqlPrismaService.matchPost.findFirst({
-      where: {
-        id: matchId,
-        deletedAt: null,
-      },
-      include: {
-        _count: {
-          select: {
-            MatchApplication: {
-              where: {
-                status: 'accepted',
-              },
-            },
-          },
-        },
-      },
-    });
-
-    if (!matchPost) {
-      throw new NotFoundException('Match post not found');
-    }
-
-    // 자신의 게시글에는 신청할 수 없음
-    if (matchPost.userno === applicantUserno) {
-      throw new BadRequestException('You cannot apply to your own match post');
-    }
-
-    // 최대 참가자 수 확인
-    if (matchPost._count.MatchApplication >= matchPost.maxParticipants) {
-      throw new BadRequestException('Match post is already full');
-    }
-
-    // 이미 신청한 사용자인지 확인
-    const existingApplication =
-      await this.mysqlPrismaService.matchApplication.findFirst({
-        where: {
-          matchPostId: matchId,
-          applicantUserno,
-        },
-      });
-
-    if (existingApplication) {
-      throw new BadRequestException(
-        'You have already applied to this match post',
-      );
-    }
-
-    await this.mysqlPrismaService.matchApplication.create({
-      data: {
-        matchPostId: matchId,
-        applicantUserno,
-        applicantName,
-        message,
-        status: 'pending',
-      },
-    });
-
-    await this.mysqlPrismaService.notification.create({
-      data: {
-        userId: matchPost.userno,
-        type: 'match_apply',
-        title: '새로운 매칭 신청',
-        body: `${applicantName}님이 "${matchPost.title}"에 신청했습니다.`,
-        targetId: matchId,
-      },
-    }).catch(() => {});
-
-    return {
-      success: true,
-      message: 'Application submitted successfully',
-    };
-  }
-
-  async getMatchApplications(
-    request: GetMatchApplicationsRequest,
-  ): Promise<MatchApplicationsResponse> {
-    const { matchId, userno } = request;
-    this.logger.debug(`getMatchApplications matchId=${matchId} requestedBy=${userno}`);
-
-    // 게시글 존재 여부 확인
-    const matchPost = await this.mysqlPrismaService.matchPost.findFirst({
-      where: {
-        id: matchId,
-        deletedAt: null,
-      },
-    });
-
-    if (!matchPost) {
-      throw new NotFoundException('Match post not found');
-    }
-
-    // 권한이 없으면 빈 배열 반환 (에러를 내지 않음)
-    if (matchPost.userno !== userno) {
-      return {
-        applications: [],
-      };
-    }
-
-    const applications =
-      await this.mysqlPrismaService.matchApplication.findMany({
-        where: {
-          matchPostId: matchId,
-        },
-        include: {
-          User: { select: { gender: true } },
-        },
-        orderBy: {
-          createdAt: 'desc',
-        },
-      });
-    this.logger.debug(`getMatchApplications matchId=${matchId} count=${applications.length}`);
-    const formattedApplications: MatchApplication[] = applications.map((app) => ({
-      id: app.id,
-      matchPostId: app.matchPostId,
-      applicantUserno: app.applicantUserno,
-      applicantName: app.applicantName,
-      message: app.message,
-      status: app.status,
-      createdAt: app.createdAt.toISOString(),
-      gender: app.User?.gender || '',
-    }));
-
-    return {
-      applications: formattedApplications,
-    };
-  }
-
-  async updateApplicationStatus(
-    request: UpdateApplicationStatusRequest,
-  ): Promise<ApplicationResponse> {
-    const { matchId, applicationId, status, userno } = request;
-    this.logger.log(`updateApplicationStatus match=${matchId} app=${applicationId} status=${status} userId=${userno}`);
-
-    // 게시글 존재 여부 및 권한 확인
-    const matchPost = await this.mysqlPrismaService.matchPost.findFirst({
-      where: {
-        id: matchId,
-        deletedAt: null,
-      },
-    });
-
-    if (!matchPost) {
-      throw new NotFoundException('Match post not found');
-    }
-
-    if (matchPost.userno !== userno) {
-      throw new ForbiddenException(
-        'You can only manage applications for your own posts',
-      );
-    }
-
-    // 신청서 존재 여부 확인
-    const application =
-      await this.mysqlPrismaService.matchApplication.findFirst({
-        where: {
-          id: applicationId,
-          matchPostId: matchId,
-        },
-      });
-
-    if (!application) {
-      throw new NotFoundException('Application not found');
-    }
-
-    // 이미 처리된 신청서인지 확인
-    if (application.status !== 'pending') {
-      throw new BadRequestException('Application has already been processed');
-    }
-
-    // 수락하는 경우 최대 참가자 수 확인
-    if (status === 'accepted') {
-      const acceptedCount =
-        await this.mysqlPrismaService.matchApplication.count({
-          where: {
-            matchPostId: matchId,
-            status: 'accepted',
-          },
-        });
-
-      if (acceptedCount >= matchPost.maxParticipants) {
-        throw new BadRequestException('Match post is already full');
-      }
-    }
-
-    // 신청서 상태 업데이트
-    await this.mysqlPrismaService.matchApplication.update({
-      where: {
-        id: applicationId,
-      },
-      data: {
-        status: status as any, // Prisma enum type 캐스팅
-      },
-    });
-
-    let chatRoomId = '';
-    let message = '';
-
-    if (status === 'accepted') {
-      // 채팅방 생성
-      chatRoomId = await this.createChatRoom(
-        matchPost.userno,
-        application.applicantUserno,
-      );
-      message = 'Application accepted successfully. Chat room created.';
-    } else if (status === 'rejected') {
-      message = 'Application rejected successfully.';
-    }
-
-    return {
-      success: true,
-      message,
-      chatRoomId,
-    };
-  }
-
-  // 채팅 서비스와 연동하여 채팅방 생성
-  private async createChatRoom(
-    authorUserno: number,
-    applicantUserno: number,
-  ): Promise<string> {
-    try {
-      const response = await firstValueFrom(
-        this.chatService.createChatRoom({
-          memberIds: [authorUserno, applicantUserno],
-          roomName: `Match Chat ${authorUserno}-${applicantUserno}`,
-          type: 'direct',
-        }),
-      );
-      return response.chatRoom.id;
-    } catch (error) {
-      this.logger.error('Failed to create chat room', error?.stack ?? error);
-      throw new BadRequestException('Failed to create chat room');
-    }
-  }
-
-  // 내가 작성한 매치 게시글 조회
-  async getMyPosts(request: GetMyPostsRequest): Promise<MatchPostResponse> {
-    const { userno, page = 1, pageSize = 10 } = request;
-    const skip = (page - 1) * pageSize;
-
-    const matchPosts = await this.mysqlPrismaService.matchPost.findMany({
-      where: {
-        userno: userno,
-        deletedAt: null,
-      },
-      include: {
-        User: true,
-        _count: {
-          select: {
-            MatchApplication: {
-              where: {
-                status: 'accepted',
-              },
-            },
-          },
-        },
-      },
-      orderBy: {
-        createdAt: 'desc',
-      },
-      skip,
-      take: pageSize + 1, // 하나 더 가져와서 hasNext 확인
-    });
-
-    const hasNext = matchPosts.length > pageSize;
-    if (hasNext) {
-      matchPosts.pop(); // 마지막 요소 제거
-    }
-
-    const formattedPosts: MatchPost[] = matchPosts.map((post) => ({
-      id: post.id,
-      title: post.title,
-      userno: post.userno,
-      author: post.User.nickname || 'Unknown',
-      authorGender: post.User.gender || '',
-      content: post.content,
-      movieTitle: post.movieTitle,
-      theaterName: post.theaterName,
-      showTime: post.showTime,
-      maxParticipants: post.maxParticipants,
-      currentParticipants: post._count.MatchApplication,
-      location: post.location,
-      createdAt: post.createdAt.toISOString(),
-      updatedAt: post.updatedAt?.toISOString() || '',
-      deletedAt: post.deletedAt?.toISOString() || '',
-    }));
-
-    return {
-      matchPosts: formattedPosts,
-      hasNext,
-    };
-  }
-
-  // 내가 신청한 매치 신청서 조회
-  async getMyApplications(
-    request: GetMyApplicationsRequest,
-  ): Promise<MatchApplicationsResponse> {
-    const { userno, page = 1, pageSize = 10 } = request;
-    const skip = (page - 1) * pageSize;
-    this.logger.debug(`getMyApplications userId=${userno}`);
-
-    const applications =
-      await this.mysqlPrismaService.matchApplication.findMany({
-        where: {
-          applicantUserno: userno,
-        },
-        include: {
-          MatchPost: {
-            include: {
-              User: true,
-            },
-          },
-          User: { select: { gender: true } },
-        },
-        orderBy: {
-          createdAt: 'desc',
-        },
-        skip,
-        take: pageSize,
-      });
-
-    const formattedApplications: MatchApplication[] = applications.map((app) => ({
-      id: app.id,
-      matchPostId: app.matchPostId,
-      applicantUserno: app.applicantUserno,
-      applicantName: app.applicantName,
-      message: app.message,
-      status: app.status,
-      createdAt: app.createdAt.toISOString(),
-      gender: app.User?.gender || '',
-    }));
-
-    return {
-      applications: formattedApplications,
-    };
-  }
-
-  // 특정 매치에 대한 내 신청 상태 조회
-  async getMyApplicationStatus(request: {
-    matchId: string;
-    userno: number;
-  }): Promise<{
-    application?: MatchApplication;
-    hasApplication: boolean;
-  }> {
-    const { matchId, userno } = request;
-
-    // 게시글 존재 여부 확인
-    const matchPost = await this.mysqlPrismaService.matchPost.findFirst({
-      where: {
-        id: matchId,
-        deletedAt: null,
-      },
-    });
-
-    if (!matchPost) {
-      throw new NotFoundException('Match post not found');
-    }
-
-    // 내 신청서 조회
-    const application =
-      await this.mysqlPrismaService.matchApplication.findFirst({
-        where: {
-          matchPostId: matchId,
-          applicantUserno: userno,
-        },
-      });
-
-    if (!application) {
-      return {
-        hasApplication: false,
-      };
-    }
-
-    // 지원자 User 정보 조회
-    const user = await this.mysqlPrismaService.user.findUnique({
-      where: { id: application.applicantUserno },
-    });
-
-    const formattedApplication: MatchApplication = {
-      id: application.id,
-      matchPostId: application.matchPostId,
-      applicantUserno: application.applicantUserno,
-      applicantName: application.applicantName,
-      message: application.message,
-      status: application.status,
-      createdAt: application.createdAt.toISOString(),
-      gender: user?.gender || '',
-    };
-
-    return {
-      application: formattedApplication,
-      hasApplication: true,
-    };
+  getMyApplicationStatus(req: { matchId: string; userno: number }) {
+    return this.applicationService.getMyApplicationStatus(req);
   }
 }
